move common functions under util/common
